Add interactive list functionality
